01-basic-chats/openai-sample.py
from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()


client = OpenAI()

response = client.responses.create(
    model="gpt-4o-mini",
    instructions="You are a coding assistant that talks like a pirate.",
    input="How do I check if a Python object is an instance of a class?",
)

print(response.output_text)
