01-basic-chats/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.38.0
google-genai==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
websockets==15.0.1
